connection: fix exceptions thrown trying to close unconnected sockets
There are places where exceptions can be thrown during connection,
these are caught and as part of cleanup, close_socket() is invariably
called.  If (and this happens) the socket has not been setup at this
point, then an unexpected error occurs and is passed to the client.
This commit attempts to close the socket iff it exists at the point of
error, allowing the client access to the actual connection exception/
error.  It also removes resetting the socket attribute, since this is
taken care of in the call to reset_values().