connection: ensure connection is actually closed when exiting context
Using the Connection object with a context manager, clients expect
exiting the context to clean up state created when entering the
context.  This commit updates Connection.__exit__() to do just that;
notifiying the server we're terminating the connection then closing
the associated socket.

Also, retypes the exceptions thrown when attempting to operate on a
closed connection from Error to ConnectionError, improving clarity
and allowing clients to better react to these cases.