connection: only commit on context exit if necessary
The current connection context manager issues either a 'commit' or a
'rollback' command based on whether there were any exceptions
thrown during the context execution.  In the case where no write
commands were issued to the server, there's no need to perform these
operations, so at best they're just a (small) overhead.  However,
in the wild we also observe a non-trivial number of cases where
reading the return value from this no-op command actually results in
an exception (select() indicates bytes are available on the read
socket, but recv() returns none) and this causes the entire context
to appear to have failed.

This commit alters the connection context manager to only issue commit/
rollback commands when there are outstanding changes when exiting.