Use kwargs
This doesn't follow the suggestion in PEP 0249 (https://www.python.org/dev/peps/pep-0249/#id40) for the connect method to utilized keyword arguments. This may seem pedantic, but in fact this will raise errors when trying to use a generic connect method. For example, if you'd like a method to connect to both MySQL and Vertica, you will get errors.