Remove calls to unused `query` function in `Connection` class
There were some calls to a nonexistent function `query` in the
`initialize_connection` function of the `Connection` class. These
appear to be artifacts from when porting the Ruby codebase.

The queries being made were related to setting a role for the current
session, as well as setting schema search path.