Fix quoting issue with client-side %s parameters
Backslash quoting was used to escape for `re.sub`, used by :var style
placeholders. This was inappropriate for % style placeholders.

This change replaces blackslash quoting with a lambda function,
sidestepping the special characters for re.sub, and leaves leaves %
style values unescaped (except for the necessary '' quoting, which
remains in place.)